ref: rename externalcontrollers -> sgreconciler
