+++ operators/constellation-node-operator/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "//3rdparty/node-maintenance-operator/api/v1beta1",
         "//operators/constellation-node-operator/api/v1alpha1",
         "//operators/constellation-node-operator/controllers",
-        "//operators/constellation-node-operator/externalcontrollers",
         "//operators/constellation-node-operator/internal/cloud/api",
         "//operators/constellation-node-operator/internal/cloud/aws/client",
         "//operators/constellation-node-operator/internal/cloud/azure/client",
@@ -23,6 +22,7 @@ go_library(
         "//operators/constellation-node-operator/internal/etcd",
         "//operators/constellation-node-operator/internal/executor",
         "//operators/constellation-node-operator/internal/upgrade",
+        "//operators/constellation-node-operator/sgreconciler",
         "@io_k8s_apimachinery//pkg/runtime",
         "@io_k8s_apimachinery//pkg/util/runtime",
         "@io_k8s_client_go//discovery",
