+++ operators/constellation-node-operator/sgreconciler/BUILD.bazel
@@ -2,30 +2,39 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 load("//bazel/go:go_test.bzl", "go_test")
 
 go_library(
-    name = "externalcontrollers",
+    name = "sgreconciler",
     srcs = [
-        "externalcontrollers.go",
+        "nodejoin_watcher.go",
         "scalinggroup_controller.go",
+        "sgreconciler.go",
     ],
-    importpath = "github.com/edgelesssys/constellation/v2/operators/constellation-node-operator/v2/externalcontrollers",
+    importpath = "github.com/edgelesssys/constellation/v2/operators/constellation-node-operator/v2/sgreconciler",
     visibility = ["//visibility:public"],
     deps = [
         "//internal/constants",
         "//operators/constellation-node-operator/api/v1alpha1",
         "//operators/constellation-node-operator/internal/cloud/api",
         "//operators/constellation-node-operator/internal/executor",
+        "@io_k8s_api//core/v1:core",
         "@io_k8s_apimachinery//pkg/api/errors",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:meta",
+        "@io_k8s_apimachinery//pkg/runtime",
+        "@io_k8s_apimachinery//pkg/types",
         "@io_k8s_client_go//util/retry",
+        "@io_k8s_sigs_controller_runtime//:controller-runtime",
+        "@io_k8s_sigs_controller_runtime//pkg/builder",
         "@io_k8s_sigs_controller_runtime//pkg/client",
+        "@io_k8s_sigs_controller_runtime//pkg/event",
+        "@io_k8s_sigs_controller_runtime//pkg/handler",
         "@io_k8s_sigs_controller_runtime//pkg/log",
+        "@io_k8s_sigs_controller_runtime//pkg/predicate",
     ],
 )
 
 go_test(
-    name = "externalcontrollers_test",
+    name = "sgreconciler_test",
     srcs = ["scalinggroup_controller_test.go"],
-    embed = [":externalcontrollers"],
+    embed = [":sgreconciler"],
     deps = [
         "//internal/constants",
         "//operators/constellation-node-operator/api/v1alpha1",
