+++ operators/constellation-node-operator/sgreconciler/nodejoin_watcher.go
@@ -4,7 +4,7 @@ Copyright (c) Edgeless Systems GmbH
 SPDX-License-Identifier: AGPL-3.0-only
 */
 
-package controllers
+package sgreconciler
 
 import (
 	"context"
