+++ operators/constellation-node-operator/sgreconciler/scalinggroup_controller.go
@@ -4,7 +4,7 @@ Copyright (c) Edgeless Systems GmbH
 SPDX-License-Identifier: AGPL-3.0-only
 */
 
-package externalcontrollers
+package sgreconciler
 
 import (
 	"context"
