+++ operators/constellation-node-operator/sgreconciler/scalinggroup_controller_env_test.go
@@ -6,7 +6,7 @@ Copyright (c) Edgeless Systems GmbH
 SPDX-License-Identifier: AGPL-3.0-only
 */
 
-package externalcontrollers
+package sgreconciler
 
 import (
 	"context"
